[BE] refactor : 필터 클래스 도메인으로 이동
