+++ backend/src/main/java/moim_today/application/moim/moim/MoimServiceImpl.java
@@ -1,7 +1,7 @@
 package moim_today.application.moim.moim;
 
 import moim_today.dto.moim.moim.*;
-import moim_today.dto.moim.moim.filter.MoimFilterRequest;
+import moim_today.dto.moim.moim.MoimFilterRequest;
 import moim_today.implement.file.FileUploader;
 import moim_today.implement.meeting.joined_meeting.JoinedMeetingRemover;
 import moim_today.implement.meeting.meeting.MeetingFinder;
