+++ backend/src/main/java/moim_today/domain/moim/MoimSortedFilter.java
@@ -1,4 +1,4 @@
-package moim_today.dto.moim.moim.filter;
+package moim_today.domain.moim;
 
 public enum MoimSortedFilter {
     CREATED_AT, VIEWS
