+++ backend/src/main/java/moim_today/dto/moim/moim/MoimFilterRequest.java
@@ -1,6 +1,7 @@
-package moim_today.dto.moim.moim.filter;
+package moim_today.dto.moim.moim;
 
 import lombok.Builder;
+import moim_today.domain.moim.MoimSortedFilter;
 import moim_today.domain.moim.enums.MoimCategory;
 
 @Builder
