+++ backend/src/main/java/moim_today/implement/moim/moim/MoimFinder.java
@@ -3,7 +3,7 @@ package moim_today.implement.moim.moim;
 import moim_today.dto.moim.moim.MoimDateResponse;
 import moim_today.dto.moim.moim.MoimMemberResponse;
 import moim_today.dto.moim.moim.MoimSimpleResponse;
-import moim_today.dto.moim.moim.filter.MoimFilterRequest;
+import moim_today.dto.moim.moim.MoimFilterRequest;
 import moim_today.global.annotation.Implement;
 import moim_today.global.error.BadRequestException;
 import moim_today.implement.member.MemberFinder;
