+++ backend/src/main/java/moim_today/persistence/repository/moim/moim/MoimRepository.java
@@ -2,7 +2,7 @@ package moim_today.persistence.repository.moim.moim;
 
 import moim_today.dto.moim.moim.MoimDateResponse;
 import moim_today.dto.moim.moim.MoimSimpleResponse;
-import moim_today.dto.moim.moim.filter.MoimFilterRequest;
+import moim_today.dto.moim.moim.MoimFilterRequest;
 import moim_today.persistence.entity.moim.moim.MoimJpaEntity;
 
 import java.util.List;
