+++ backend/src/main/java/moim_today/persistence/repository/moim/moim/MoimRepositoryImpl.java
@@ -9,8 +9,8 @@ import moim_today.dto.moim.moim.MoimDateResponse;
 import moim_today.dto.moim.moim.MoimSimpleResponse;
 import moim_today.dto.moim.moim.QMoimDateResponse;
 import moim_today.dto.moim.moim.QMoimSimpleResponse;
-import moim_today.dto.moim.moim.filter.MoimFilterRequest;
-import moim_today.dto.moim.moim.filter.MoimSortedFilter;
+import moim_today.dto.moim.moim.MoimFilterRequest;
+import moim_today.domain.moim.MoimSortedFilter;
 import moim_today.global.error.NotFoundException;
 import moim_today.persistence.entity.moim.moim.MoimJpaEntity;
 import org.springframework.stereotype.Repository;
