+++ backend/src/test/java/moim_today/presentation/moim/MoimControllerTest.java
@@ -5,13 +5,9 @@ import moim_today.application.moim.moim.MoimService;
 import moim_today.application.moim.moim_notice.MoimNoticeService;
 import moim_today.domain.moim.DisplayStatus;
 import moim_today.domain.moim.enums.MoimCategory;
-import moim_today.dto.moim.moim.MoimCreateRequest;
-import moim_today.dto.moim.moim.MoimDeleteRequest;
-import moim_today.dto.moim.moim.MoimMemberDeleteRequest;
-import moim_today.dto.moim.moim.MoimUpdateRequest;
+import moim_today.dto.moim.moim.*;
 import moim_today.dto.moim.moim.filter.MoimFilterRequest;
 import moim_today.dto.moim.moim.filter.MoimSortedFilter;
-import moim_today.dto.moim.moim.*;
 import moim_today.dto.moim.moim_notice.MoimNoticeCreateRequest;
 import moim_today.dto.moim.moim_notice.MoimNoticeDeleteRequest;
 import moim_today.dto.moim.moim_notice.MoimNoticeUpdateRequest;
